[feat]:[pl-28768]: add file path validation in pipeline, input-set and template service
